+++ lib/src/circle.dart
@@ -19,16 +19,22 @@ class Circle extends IterableBase<Vec> {
     if (radius < 0) throw ArgumentError("The radius cannot be negative.");
   }
 
-  /// Gets whether [pos] is in the outermost edge of the circle.
-  bool isEdge(Vec pos) {
-    var leadingEdge = true;
-    if (radius > 0) {
-      leadingEdge = (pos - center) > _radiusSquared(radius - 1);
-    }
+  /// Whether [pos] is within the circle.
+  @override
+  bool contains(covariant Vec pos) => _contains(center, radius, pos);
 
-    return leadingEdge;
+  /// Whether [pos] is in the outermost edge of the circle.
+  bool isEdge(Vec pos) {
+    // Must be within the circle.
+    if (!contains(pos)) return false;
+
+    // But not within the next smallest circle inside it.
+    if (radius > 0 && _contains(center, radius - 1, pos)) return false;
+
+    return true;
   }
 
+  /// Iterates over all points within the circle.
   @override
   Iterator<Vec> get iterator => _CircleIterator(this, edge: false);
 
@@ -45,15 +51,21 @@ class _CircleIterable extends IterableBase<Vec> {
 
 class _CircleIterator implements Iterator<Vec> {
   @override
-  Vec get current => _boundsIterator.current + _circle.center;
+  Vec get current => _boundsIterator.current;
 
   final Circle _circle;
+
+  /// Iterator over the rectangle of points containing the circle.
   final Iterator<Vec> _boundsIterator;
+
+  /// Whether we are iterating over just the edge points or all points within
+  /// the circle.
   final bool _edge;
 
   factory _CircleIterator(Circle circle, {required bool edge}) {
     var size = circle.radius + circle.radius + 1;
-    var bounds = Rect(-circle.radius, -circle.radius, size, size);
+    var bounds = Rect(circle.center.x - circle.radius,
+        circle.center.y - circle.radius, size, size);
     return _CircleIterator._(circle, bounds.iterator, edge: edge);
   }
 
@@ -62,27 +74,70 @@ class _CircleIterator implements Iterator<Vec> {
 
   @override
   bool moveNext() {
-    while (true) {
-      if (!_boundsIterator.moveNext()) return false;
-      var length = _boundsIterator.current.lengthSquared;
-
-      if (length > _radiusSquared(_circle.radius)) continue;
-      if (_edge &&
-          _circle.radius > 0 &&
-          length < _radiusSquared(_circle.radius - 1)) continue;
-
-      break;
+    while (_boundsIterator.moveNext()) {
+      // Skip over points that aren't within the what we're iterating over.
+      var pos = _boundsIterator.current;
+      if (_edge) {
+        if (_circle.isEdge(pos)) return true;
+      } else {
+        if (_circle.contains(pos)) return true;
+      }
     }
 
-    return true;
+    return false;
   }
 }
 
-const _radiiSquared = [0, 2, 5, 10, 18, 26, 38];
+bool _contains(Vec center, int radius, Vec pos) {
+  var lengthSquared = (pos - center).lengthSquared;
 
-int _radiusSquared(int radius) {
   // If small enough, use the tuned radius to look best.
-  if (radius < _radiiSquared.length) return _radiiSquared[radius];
+  const radiiSquared = [0, 2, 5, 10, 18, 26, 38];
+  if (radius < radiiSquared.length) {
+    return lengthSquared <= radiiSquared[radius];
+  }
 
-  return radius * radius;
+  // Otherwise, sort of split the difference between the actual squared radius
+  // and the radius of a 1-unit larger circle. If we use the actual radius, then
+  // the circle only contains the outermost points right at the origins, like:
+  //
+  // .................
+  // ........X........
+  // .....*******.....
+  // ....*********....
+  // ...***********...
+  // ..*************..
+  // ..*************..
+  // ..*************..
+  // .X*************X.
+  // ..*************..
+  // ..*************..
+  // ..*************..
+  // ...***********...
+  // ....*********....
+  // .....*******.....
+  // ........X........
+  // .................
+  //
+  // By cheating out the radius a bit, we get circles that fill their radius a
+  // little better, like:
+  //
+  // .................
+  // ......*****......
+  // ....*********....
+  // ...***********...
+  // ..*************..
+  // ..*************..
+  // .***************.
+  // .***************.
+  // .***************.
+  // .***************.
+  // .***************.
+  // ..*************..
+  // ..*************..
+  // ...***********...
+  // ....*********....
+  // ......*****......
+  // .................
+  return lengthSquared <= radius * (radius + 1);
 }